ocr-app/backend/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
python-dotenv
python-multipart
